+++ MinecraftJars.Extension/MinecraftJars.Extension.DependencyInjection/ServiceCollectionExtensions.cs
@@ -1,5 +1,4 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
-using MinecraftJars.Core.Providers;
 
 namespace MinecraftJars.Extension.DependencyInjection;
 
@@ -8,7 +7,7 @@ public static class ServiceCollectionExtensions
     public static IServiceCollection AddMinecraftJar(this IServiceCollection services)
     {
         services.AddHttpClient();
-        services.AddScoped<IMinecraftJar, MinecraftJar>(sp => new MinecraftJar(new ProviderOptions
+        services.AddScoped<IMinecraftJar, MinecraftJar>(sp => new MinecraftJar(new MinecraftJarOptions()
         {
             HttpClientFactory = sp.GetRequiredService<IHttpClientFactory>()
         }));