+++ MinecraftJars/MinecraftJarOptions.cs
@@ -1,10 +1,10 @@
-﻿namespace MinecraftJars.Core.Providers;
+namespace MinecraftJars;
 
-public class ProviderOptions
+public class MinecraftJarOptions
 {
     /// <summary>
     /// If provided the CreateClient Method is utilized to create a HttpClient
     /// otherwise a new HttpClient is instantiated by the MinecraftJarManager
     /// </summary>     
-    public IHttpClientFactory? HttpClientFactory { get; init; }
+    public IHttpClientFactory? HttpClientFactory { get; init; }    
 }